refactor: ordertable 패키지 order에서 분리
