+++ src/main/java/kitchenpos/order/Order.java
@@ -9,6 +9,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.ManyToOne;
+import kitchenpos.ordertable.OrderTable;
 
 
 @Entity(name = "orders")
