+++ src/main/java/kitchenpos/ordertable/application/OrderTableService.java
@@ -1,11 +1,11 @@
-package kitchenpos.order.application;
+package kitchenpos.ordertable.application;
 
 import java.util.List;
 import kitchenpos.order.OrderStatus;
-import kitchenpos.order.OrderTable;
 import kitchenpos.order.persistence.OrderRepository;
-import kitchenpos.order.persistence.OrderTableRepository;
-import kitchenpos.order.request.OrderTableCreateRequest;
+import kitchenpos.ordertable.OrderTable;
+import kitchenpos.ordertable.persistence.OrderTableRepository;
+import kitchenpos.ordertable.request.OrderTableCreateRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
