+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -1,16 +1,16 @@
-package kitchenpos.table_group.application;
+package kitchenpos.tablegroup.application;
 
 import static java.util.stream.Collectors.toList;
 
 import java.util.List;
 import kitchenpos.order.Order;
-import kitchenpos.order.OrderTable;
 import kitchenpos.order.persistence.OrderRepository;
-import kitchenpos.order.persistence.OrderTableRepository;
-import kitchenpos.order.request.OrderTableDto;
-import kitchenpos.table_group.TableGroup;
-import kitchenpos.table_group.persistence.TableGroupRepository;
-import kitchenpos.table_group.request.TableGroupCreateRequest;
+import kitchenpos.ordertable.OrderTable;
+import kitchenpos.ordertable.persistence.OrderTableRepository;
+import kitchenpos.tablegroup.TableGroup;
+import kitchenpos.tablegroup.persistence.TableGroupRepository;
+import kitchenpos.tablegroup.request.OrderTableDto;
+import kitchenpos.tablegroup.request.TableGroupCreateRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
