+++ src/main/java/kitchenpos/tablegroup/persistence/TableGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.table_group.persistence;
+package kitchenpos.tablegroup.persistence;
 
-import kitchenpos.table_group.TableGroup;
+import kitchenpos.tablegroup.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
