+++ src/main/java/kitchenpos/ui/TableGroupRestController.java
@@ -1,9 +1,9 @@
 package kitchenpos.ui;
 
 import java.net.URI;
-import kitchenpos.table_group.TableGroup;
-import kitchenpos.table_group.application.TableGroupService;
-import kitchenpos.table_group.request.TableGroupCreateRequest;
+import kitchenpos.tablegroup.TableGroup;
+import kitchenpos.tablegroup.application.TableGroupService;
+import kitchenpos.tablegroup.request.TableGroupCreateRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
