+++ src/test/java/kitchenpos/order/application/OrderTableServiceTest.java
@@ -10,13 +10,14 @@ import java.util.List;
 import kitchenpos.order.Order;
 import kitchenpos.order.OrderLineItem;
 import kitchenpos.order.OrderStatus;
-import kitchenpos.order.OrderTable;
 import kitchenpos.order.persistence.OrderRepository;
-import kitchenpos.order.persistence.OrderTableRepository;
-import kitchenpos.order.request.OrderTableCreateRequest;
+import kitchenpos.ordertable.OrderTable;
+import kitchenpos.ordertable.application.OrderTableService;
+import kitchenpos.ordertable.persistence.OrderTableRepository;
+import kitchenpos.ordertable.request.OrderTableCreateRequest;
 import kitchenpos.support.ServiceTest;
-import kitchenpos.table_group.TableGroup;
-import kitchenpos.table_group.persistence.TableGroupRepository;
+import kitchenpos.tablegroup.TableGroup;
+import kitchenpos.tablegroup.persistence.TableGroupRepository;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Nested;
